old-tests/webapps/WebStorage/tests/approved/event_session_url.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>event_session_url</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		var t = async_test("url property test of session event");

		function onStorageEvent(event) {
			var url = window.location.href;

			var pos = url.lastIndexOf("/");
			if (pos != -1) {
				url = url.substr(0, pos + 1);
				url = url + "iframe/session_set_item_iframe.html";
			}

			t.step(function() {
			   assert_equals(event.url, url);
			  });
			t.done();
		}

		if (window.addEventListener) {
            window.addEventListener('storage', onStorageEvent, false);
        }

        else {
			test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			}, "existence of event listener")
        }

		var el = document.createElement("iframe");
		el.setAttribute('id', 'ifrm');
		el.setAttribute('src', 'iframe/session_set_item_iframe.html');
		el.style.border='0px';
		el.style.width='0px';
		el.style.height='0px'		
		document.body.appendChild(el);

	</script>

 </body>
</html>
